Avoid SEGV during backend initialization due to elog(ERROR).
This is fixed in PostgreSQL and I found parts of those fixes pulled in already.
This commit pulls in the remnants.  The SEGV was found at a customer site
where default_tablespace was configured for a role.  While new connections were
made by the role, GRANT/REVOKE on the same tablespace was running concurrently.
The assign hook for default_tablespace invoked elog(ERROR) but target for
siglongjmp was not initialized, causing a segv and PANIC.

Some unit tests had to be changed to accommodate the fix.

Original commits from PostgreSQL:
88f1fd29897df477f0af3c5ffcefe53c697a6ff3
79ca7ffeb6a43c1f2af80ead6a2dffd120594cf9