DECLARE CURSOR don't error out during dispatch.
In GPDB, QEs start execution once the plan of DECLARE CURSOR is dispatched, and the
errors of execution will be reported back to dispatcher. As threads are scheduled
by OS, the execution sequence of multiple dispatcher threads is unpredictable, which
will make the behavior of DECLARE CURSOR inconsistent.
This fix keeps DECLARE CURSOR from erroring out in dispatch phase and the errors will
be reported during FETCH or at the end of the transaction.