Remove the shared memory object only when transaction of drop resource group is committed
Previously, we remove the shared memory object when drop resource group,
and restore it if the transaction aborts. The concurrently access to
shared memory object would fail in this way.
If the resource group is dropped, new transactions to this resource
group will be queued up until the drop transaction is finished.

Signed-off-by: Richard Guo <riguo@pivotal.io>