resgroup: Fix memory quota calculation
If a QD exits when the transaction is still active, QD will abort
the transaction and destroy all gangs, without waiting for QEs
finish the transaction. In this case, a new transaction (the resource
group slot waken up by this QD) may execute before the QEs exit,
so we cannot use the number of currently running QEs to calculate
the memory quota.