Add hacks to keep track of EOS
So, this doesn't work. I tried to add in the pseudo-code to actually
have the receiver send an EOS Ack if it got an end of stream and had no
new params and to have the sender check for EOSAcks before finishing.

Also, I noticed that the estate expression context ecxt_param_exec_vals
in the motion sender slice is uninitialized at the point of
execMotionSender. It only gets initialized in ExecInitMotion when there
are a number of params more than 0 in the plan state. I decided that,
for now, I would manually, initialize it in execMotionSender and then
later figure out how to get the number of params from the receiver and
copy it over. However, I have a feeling that my initializaiton in the
motionstate planstate alone will be insufficient as soon as I stop
hard-coding the scanvalue when evaluating the indexscan.

I hard-coded the number of rescans (num params) to be 5 in the motion
sender.

When I execute this code, I get an error:
ERROR:  cannot execute squelched plan node of type: 229
(execProcnode.c:951)  (seg2 slice2 127.0.0.1:25434 pid=31007)

I'm not sure why, but it looks like the NLJ is never producing tuples so
it gets squelched. Something is wrong with my code.

Also, I thought about when to match EOS and param. Probably receiver
needs to check if EOS is for the current param it is sending before
sending a new param (I added some comment with a description of that).
However, I don't know how it will know which EOS is the EOS it got in
RecvTuple, because, currently, I can't see the packet that had the EOS
from the caller of RecvTuple. Maybe I should use my super hack-y
pass-by-reference isEOS bool to actually be the seq number of the packet
so I can compare it to the seq number of the param packet I sent.
Though, there again, I don't send the actual packet in the
motionReceiver code, I call a thing that does that