Super broken try and match up sequences
So, this hangs for most queries and makes it so you can't even cancel
the query--pretty sure it hangs the QD. The logic is super broken, but I
am committing it for now. Feel free to completely delete all of this, as
the code before this commit without the logic to check sequences doesn't
work either, so, this is pretty premature.