Don't put EOSseq in chunk sorter
I put the EOSseq in the totally wrong place (chunk sorter is for sorted
motion receiver, which I didn't even try to change anyway). I added it
as a field to the motion state node and use it for motion receivers. It
is still not right. Basically, now, if all the senders have sent EOS, I
consider the reciever to have gotten EOS and then I check if the EOSSeq
is the same as the paramseq before I consider it the correct EOS.

I think I will roll back to the commit before I tried adding the
sequences and try to fix that first.