Remove code to poll acks and params that doesn't work
So, pollParams and pollEOSAcks were not necessary and did not work. Previously,
I thought that execMotionSender could check for params being sent and EOS Acks
being sent. this doesn't make sense because the sender This doesn't work at all
because there are two distinct phases for the sender: 1 is sending data, at
which time, it sends tuples and polls for acks in case it should stop sending
the tuples It doesn't stop sending until it has gotten all acks Then it
proceeds to EOS phase in which it sends EOS and then waits for EOS ack messages
from the receivers Now, I've added the code here so that if it gets a param
packet (instead of an ack packet) while it is in the EOS phase, then it will
know it needs to initiate rescan

The database still crashes with this code. Other problems exist.

One outstanding item is that the receiver needs to notify the interconnect
thread when it should send EOS Ack to the sender, based on whether or not it
has more params to send. This is not done at all, so the receiver will send EOS
Ack happily back to the sender for any EOS it receives and to instead sent the
next param when it gets an EOS

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>