Hack set param in child's econtext for sender and receiver
so, in the estate, we have an array of expression contexts--for each
node (I think). So, even though we initialize each node's econtext to
point to the estate's econtext for that node, we have to make sure that
we actually set the param value in the estate's econtext for that node.

I tried to find an example of a plan which had a direct inner child of
the NLJ which was not a scan or material node so that I could see how
the econtext for that specific node got its param value set in the
expression context ecxt_param_exec_vals, but I couldn't get a plan like
that which was still simple enough to understand. For a plan with
initPlan or subPlans, there are some functions that we call that might
explain how a NLJ, for example, propagates the changed parameter down to
its inner child when its inner child is not a scan.
The functions ExecReScanSetParamPlan and UpdateChangedParamSet stick out
in particular.

In the absence of a better solution, I tried simply hacking in the param
to the child of the NLJ when it is a motion (receiver side). And then
doing the same on the sender side using the hasNewParam variable to
determine when to do this.

I planned to make this a temporary hack and replace it with something
better once I got the right answer. However, a weird thing was happening
which I decided is better to fix before making the fix better.
So, for the first few values of the outer side, we don't actually get
those to the indexscan for some reason. So, if you run the indexscan
example, you will get wrong results and will miss any values before 5. I
added a logging message to log the scanvalues that we actually see. If
you insert more data that will be passed as a param later
(for me it works to simply add more data later after the initial insert)
and will meet the join condition, it will actually get returned.

So, if the first few params are being lost, I thought maybe this is an
issue with syncrhonization and not locking the param value in the
interconnect control data structure--but I have no evidence of this.