Remove hard-coded expectation in sender that child is indexonlyscan
Added a small function to check if the child is indexonly, index, or seq
scan in sender motion node. It errors out if it is not. I need to
understand what cases you would have a sender motion with a child that
it is trying to set a parameter in that is not a scan, and, if we need
it for dynamic*scans.