Optimize 'fastIUD' transactions
'fastIUD' transaction refers to an implicit transaction that does
INSERT/UPDATE/DELETE and only affects rows in one segment. We can simplify the
two-phase commit by treating it as a local transaction on the segment, which
is, QD does the planning and dispatch the plan to QE, QE execute the plan in a
local transaction context. The modification to the tuples is visible once the
transaction is committed locally. As the QD doesn't start a distributed
transaction, it's not appeared in the distributed snapshot of any other
transactions. Other transactions will check the visibility of the affected
tuples use local snapshot on the QE.