Keep the order of reusing idle gangs
For example:
In the same session,
query 1 has 3 slices and it creates gang 1, gang 2 and gang 3.
query 2 has 2 slices, we hope it reuses gang 1 and gang 2 instead of other
cases like gang 3 and gang 2.

In this way, the two queries can have the same send-receive port pair. It's
useful in platform like Azure. Because Azure limits the number of different
send-receive port pairs (AKA flow) in a certain time period.

Co-authored-by: Ning Yu <nyu@pivotal.io>
Co-authored-by: Hubert Zhang <hzhang@pivotal.io>