Remove currentDistribXid
Read-only transactions don't necessarily need distributed transaction IDs, but
some components (PXF/gpfdist/GDD) are assuming there is always a distributed
transaction ID.
We used to generate a transaction ID and save it in global variable
'currentDistribXid', assign to MyTmGxact->gxid when it's identified as a
two-phase commit transaction. It turns out it's easier to set MyTmGxact->gxid
when a transaction starts.