Optimize one-phase commit
We assign distributed transaction ID for every transaction: 1PC transaction,
2PC transaction, and implicit auto-commit transaction. The consequence is we
have to clear the transaction ID (MyTmGxact->gxid) when the transaction ends
and currently we need acquire ProcArrayLock to clear it, which is a huge lock
contention on the master running OLTP workloads. The reason for taking
ProcArrayLock is we need 'strict serialization of commits and rollbacks with
snapshot-taking: we do not allow any transaction to exit the set of running
transactions while a snapshot is being taken'.

The truth is we don't actually need to consider one-phase commit transactions
when taking a distributed snapshot. In this commit, we introduce a flag
'MyTmGxact->includeInSnapshot' to specify should the snapshot-taking process
peek this transaction's gxid. The flag is set right before dispatch 'PREPARE'
command (bypass one-phase commit), and it is cleared with ProcArrayLock held to
make sure the serialization with snapshot-taking.