Change 'latestWriteXid' from distributed transaction id to local transaction id
'latestWriteXid' is used to pick the correct segno of ao table and it's
currently a distributed transaction id, the problem is we don't always have the
distributed transaction id generated at the time. There's no reason we cannot
use the local transaction id instead as 'latestWriteXid' is updated and used on
QD only.