Fix inconsistent snapshot during concurrent update
After enabling the global deadlock detector, we can support concurrent updates.
When updating one tuple at the same time, the conflict and wait are moved from
QD to QE. We need to make sure the implicated transaction order on the segments
is also considered on the master when taking the distributed snapshots.

This is reported: https://github.com/greenplum-db/gpdb/issues/9407