Handle concurrent update 'distributed oldestXmin'
The distributed oldestXmin (DistributedLogShared->oldestXmin) is caculated with
xminAllDistributedSnapshots (in distributed snapshot) and the local oldestXmin,
each backend should has its own. However, we need a valid initial value for a
new spawn backend, so we make it shared across backends on one segment. It
won't jeopardize the correctness, because when a backend use a oldestXmin
beyond its xminAllDistributedSnapshots, the distributed transaction id of the
oldestXmin will still smaller than the xmin of its distributed snapshot.

As the backends share the oldestXmin, it can be updated concurrently, we need to
make sure we don't overwritten the value with a smaller one.

Besides, SimpleLruReadPage may release the lock, truncate the distributed log when
advancing the oldestXmin may cause conflicts. So change the timing of truncate file.