Refactor work_set cleanup
work_set could live in a transactions or through transactions (e.g. with hold
cursor). We used to cleanup the work_set when at the time proc exits, in this
case, we might leave some work_set dangling if the work_set has no work file
registered to it. This commit cleanup the transactional work_set at the end of
the transaction by traversing 'workfile_shared->activeList' and cleanup the
non-transactional work_set when the proc exits. It should resolve the issues
that #10186 and #10240 want to fix.