Truncate distributed log when vacuuming database
Truncate distributed log at the time of advancing oldest xmin can lead to heavey
lock contention on 'DistributedLogControlLock' and 'DistributedLogTruncateLock'.
It's safe to truncate the distributed log to the frozen xid when vacuuming the
database and hence can get rid of 'DistributedLogTruncateLock'. This commit also
truncate the distributed log and bump 'DistributedLogShared->oldestXmin' to the
oldest active xid when startup.